PIM-6784: Fix identifier filter
Context:

Before this PR, the identifier filter would only use the es document
property "identifier" to filter on product and product models
identifier.

This means that whenever you would filter on an identifier attribute
like "Sku", this filter would only use the "identifier" property of
the ES document to find the right node to return.

As mentionned above, the product model codes are also indexed in
the "identifier" property of the ES document.
The consequence of that is, if you are searching for a document in
ES which has an identifier equals to "my_identifier", ES would indeed
match the product model having this a code "my_identifier" as well as
the product having a sku "my_identifier".

Still, this way of searching was functioning very well in the datagrid (smart search or aggregated search) since along with
the identifier to search we would tell elasticsearch to match only the documents that would have the "sku" attribute in the document level.

This last clause would disqualify all product models from appearing in the search (which is what we want when you look for a not having the sku "toto" for instance).

Problem:
The new feature is implemented in such a way that we don't look if the
attributes are in the level of the document we match. Instead, we
just see if the node has the attribute in his ancestors (if not,
therefore he should have it, more on that in this PR:
https://github.com/akeneo/pim-community-dev/pull/7523)

Long story short: the product models are not disqualified anymore when
we filter on the attribute sku as it is implemented.

Solution: Stop relying on the "identifier" property when we search on
attribute identifiers like "sku"

1. Make sure the product values identifiers are indexed in the "values"
of the ES documents.
2. When using the PQB:

```
$pqb->addFilter('identifier', Operators:Equals, 'toto') => will use the
"identifier" property of the ES document

$pqb->addFilter('sku', Operators:Equals, 'toto') => will use the
indexed product value stored in
values.sku-text.<all_channels>.<all_locales> property of the ES
document.
```